Resolved issue related to leftover files.
If Viper was not shut down before a reboot, the id and pid files would
persist. If viper were then started, it would find these files and
attempt to kill whatever process existed at that PID. (Not good). This
change moves these files to /tmp/viper4linux so they will be removed on
a reboot.